# Marginally optimize `subdivide_grid()` in `collision.py`.