- Nested loop in `subdivide_grid()` have been swapped to allow for pre-calculating of expensive variables (`box_mid_x`, and `box_mid_z`), that are needed for each triangle test.
- Conditional checks in `collide()` have been reordered to early out as soon as it is known that there is no collision, avoiding relatively abundant calls to `min()` and `max()`, as well as offsets in tuples.

Some numbers when loading a BMD file (stock Rainbow Road):

_**Before the patch**_
```
Elapsed time: 3.0945558547973633 seconds
Elapsed time: 3.062800168991089 seconds
Elapsed time: 3.1031606197357178 seconds
Elapsed time: 3.0819194316864014 seconds
Elapsed time: 3.1193764209747314 seconds
Elapsed time: 3.0952560901641846 seconds
```

_**After the patch**_
```
Elapsed time: 2.6088180541992188 seconds
Elapsed time: 2.6334593296051025 seconds
Elapsed time: 2.6266448497772217 seconds
Elapsed time: 2.638592481613159 seconds
Elapsed time: 2.633305788040161 seconds
Elapsed time: 2.6060266494750977 seconds
```

On average, time has been reduced by `400 ms`.